Splitted tests to file based.
