+++ test/api/v1/users/get.js
@@ -1,0 +1,26 @@
+'use strict';
+
+const Lab = require('lab');
+const Code = require('code');
+
+const lab = exports.lab = Lab.script();
+const describe = lab.describe;
+const it = lab.it;
+const expect = Code.expect;
+
+const superagent = require('superagent');
+const API_URL = 'localhost:8081';
+
+describe('/api/v1/users endpoints', () => {
+  describe('GET /api/v1/users', () => {
+    it('handle naked request', (done) => {
+      superagent
+        .get(API_URL + '/api/v1/users')
+        .end((err, res) => {
+          expect(res.statusCode).to.equal(200);
+          done();
+      });
+    });
+  });
+});
+
